requirements.txt
# Core
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Psycopg2-binary==2.9.9

# Seguridad
Flask-Security-Too==5.3.1
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
Flask-Cors==4.0.0
Itsdangerous==2.1.2

# Utilidades
python-dotenv==1.0.0
Pillow==10.1.0
Markdown==3.5.1
Bleach==6.1.0
WeasyPrint==60.2
Flask-Mail==0.9.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Frontend
tailwind==3.1.0